fix spelling: contraint -> constraint
