+++ opentasks/src/main/java/org/dmfs/tasks/model/constraints/NotAfter.java
@@ -15,7 +15,7 @@
  * 
  */
 
-package org.dmfs.tasks.model.contraints;
+package org.dmfs.tasks.model.constraints;
 
 import org.dmfs.tasks.model.ContentSet;
 import org.dmfs.tasks.model.adapters.TimeFieldAdapter;
@@ -24,24 +24,16 @@ import android.text.format.Time;
 
 
 /**
- * Shift a reference time by the same amount that value has been shifted.
- * 
- * TODO: use Duration class to get the duration in days and shift without summer/winter time switches
+ * Ensure a time is not after a specific reference time. The new value will be set to the reference time otherwise.
  * 
  * @author Marten Gajda <marten@dmfs.org>
  */
-public class ShiftTime extends AbstractConstraint<Time>
+public class NotAfter extends AbstractConstraint<Time>
 {
 	private final TimeFieldAdapter mTimeAdapter;
 
 
-	/**
-	 * Creates a new ShiftTime instance.
-	 * 
-	 * @param adapter
-	 *            A {@link TimeFieldAdapter} that knows how to load the value to shift.
-	 */
-	public ShiftTime(TimeFieldAdapter adapter)
+	public NotAfter(TimeFieldAdapter adapter)
 	{
 		mTimeAdapter = adapter;
 	}
@@ -50,18 +42,13 @@ public class ShiftTime extends AbstractConstraint<Time>
 	@Override
 	public Time apply(ContentSet currentValues, Time oldValue, Time newValue)
 	{
-		Time timeToShift = mTimeAdapter.get(currentValues);
-		if (timeToShift != null && newValue != null && oldValue != null)
+		Time notAfterThisTime = mTimeAdapter.get(currentValues);
+		if (notAfterThisTime != null && newValue != null)
 		{
-			boolean isAllDay = timeToShift.allDay;
-			timeToShift.set(timeToShift.toMillis(false) + (newValue.toMillis(false) - oldValue.toMillis(false)));
-
-			// ensure the event is still allday if is was allday before.
-			if (isAllDay)
+			if (newValue.after(notAfterThisTime))
 			{
-				timeToShift.set(timeToShift.monthDay, timeToShift.month, timeToShift.year);
+				newValue.set(notAfterThisTime);
 			}
-			mTimeAdapter.set(currentValues, timeToShift);
 		}
 		return newValue;
 	}
